% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.TraitTrain.delta_00.R
\name{Simulate.TraitTrain.delta_00}
\alias{Simulate.TraitTrain.delta_00}
\title{Simulate.TraitTrain.delta_00: Function to simulate trait data under the delta model}
\usage{
Simulate.TraitTrain.delta_00(
  handle.Phylogeny,
  numeric.Sig2,
  numeric.delta,
  numeric.AncestralState,
  matrix.R
)
}
\arguments{
\item{handle.Phylogeny}{Phylogeny used to simulate training data under delta model}

\item{numeric.Sig2}{Numeric Value of the sigsg (evolutionary rate) parmater of the delta model}

\item{numeric.delta}{Numeric value of the delta parameter of the delta model}

\item{numeric.AncestralState}{Numerical value of the z0 (ancestral state) parameter of the delta model}

\item{matrix.R}{Matrix specifying the among trait covariance for p traits. Can also be a single value (1) for a single trait}

\item{numeric.NumberOfSpecies}{Numeric counting the number of species}
}
\value{
matrix.TraitData Matrix of simulated trait data. Species are rows, traits are columns
}
\description{
This function returns a Matrix of simulated trait data. Species are rows, traits are columns in matrix.TraitData
}
\examples{
#####################
# load dependencies #
#####################
library(geiger); library(phytools); library(TraitTrainR)

#########################
# get example phylogeny #
#########################
handle.PrimatePhylogeny <- read.tree(text = "((((((((human: 6, chimp:6): 1, gorilla: 7): 7, orangutan: 14): 11, macaque: 25): 64, mouse: 89): 91, opossum: 180): 20, platypus: 200): 110, chicken: 310);")

#############################
# Simulation Model Settings #
#############################
list.SimulationModelSettings <- list() # define an empty model list

##########################
# SET SIMULATION MODEL 1 #
##########################
numeric.NumberTrainingReps <- 10 # same number of replicates for all models in list.SimulationModelSettings
list.Rmatrix <- list(); for (i in 1:numeric.NumberTrainingReps){list.Rmatrix[[i]] <- matrix(c(1, 0, 0, 0, 1, 0, 0, 0, 1), nrow = 3, ncol = 3)} # three traits. Different rates for different traits can be specified here. 
#list.Rmatrix <- list(); for (i in 1:numeric.NumberTrainingReps){list.Rmatrix[[i]] <- 1} # single trait

#############################
# append to simulation list #
#############################
list.SimulationModelSettings[[1]] <- list(string.SimulationModel = "delta", 
                                          vector.Sig2 = rexp(n = numeric.NumberTrainingReps, rate = 1), 
                                          vector.AncestralState = runif(n = numeric.NumberTrainingReps, min = 0, max = 100), 
                                          vector.delta = runif(n = numeric.NumberTrainingReps, min = exp(-500), max = 3), 
                                          list.Rmatrix = list.Rmatrix)

##############################
# simulate under delta model #
##############################
Simulate.TraitTrain.delta_00(handle.Phylogeny = handle.PrimatePhylogeny, 
                             numeric.Sig2 = list.SimulationModelSettings[[1]]$vector.Sig2[1], 
                             numeric.AncestralState = list.SimulationModelSettings[[1]]$vector.AncestralState[1], 
                             numeric.delta= list.SimulationModelSettings[[1]]$vector.delta[1],
                             matrix.R = list.SimulationModelSettings[[1]]$list.Rmatrix[[1]])
}
