% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.TraitTrain.depth_STACKED_01.R
\name{Simulate.TraitTrain.depth_STACKED_01}
\alias{Simulate.TraitTrain.depth_STACKED_01}
\title{Simulate.TraitTrain.depth_STACKED_01: Function to simulate trait data under the STACK (+AncShifted) depth model}
\usage{
Simulate.TraitTrain.depth_STACKED_01(
  numeric.NumberOfSpecies,
  handle.Phylogeny,
  numeric.depth,
  numeric.AncestralState,
  vector.STACK_AncShiftNode,
  vector.STACK_AncShiftValue,
  matrix.R
)
}
\arguments{
\item{numeric.NumberOfSpecies}{Numeric counting the number of species}

\item{handle.Phylogeny}{Phylogeny used to simulate training data under depth model}

\item{numeric.AncestralState}{Numerical value of the z0 (ancestral state) parameter of the depth model}

\item{vector.STACK_AncShiftNode}{Vector that includes all nodes for a STACKED AncShift model (on top of depth)}

\item{vector.STACK_AncShiftValue}{Vector that includes all shift values for each node in vector.STACK_AncShiftNode for a STACKED AncShift model (on top of depth)}

\item{matrix.R}{Matrix specifying the among trait covariance for p traits. Can also be a single value (1) for a single trait}
}
\value{
matrix.TraitData Matrix of simulated trait data. Species are rows, traits are columns
}
\description{
This function returns a Matrix of simulated trait data. Species are rows, traits are columns in matrix.TraitData
}
\examples{
#####################
# load dependencies #
#####################
library(geiger); library(phytools); library(TraitTrainR)

#########################
# get example phylogeny #
#########################
handle.PrimatePhylogeny <- read.tree(text = "((((((((human: 6, chimp:6): 1, gorilla: 7): 7, orangutan: 14): 11, macaque: 25): 64, mouse: 89): 91, opossum: 180): 20, platypus: 200): 110, chicken: 310);")

#############################
# Simulation Model Settings #
#############################
list.SimulationModelSettings <- list() # define an empty model list

##########################
# SET SIMULATION MODEL 1 #
##########################
numeric.NumberTrainingReps <- 10 # same number of replicates for all models in list.SimulationModelSettings
matrix.STACK_AncShiftNode <- replicate(n = numeric.NumberTrainingReps, sample(size = 2, x = 1:16, replace = F))
matrix.STACK_AncShiftValue <- replicate(n = numeric.NumberTrainingReps , rnorm(n = 2, mean = 10000))
list.Rmatrix <- list(); for (i in 1:numeric.NumberTrainingReps){list.Rmatrix[[i]] <- matrix(c(1, 0, 0, 0, 1, 0, 0, 0, 1), nrow = 3, ncol = 3)} # three traits. Different rates for different traits can be specified here. 
#list.Rmatrix <- list(); for (i in 1:numeric.NumberTrainingReps){list.Rmatrix[[i]] <- 1} # single trait

#############################
# append to simulation list #
#############################
list.SimulationModelSettings[[1]] <- list(string.SimulationModel = "depth", 
                                          vector.AncestralState = runif(n = numeric.NumberTrainingReps, min = 0, max = 100), 
                                          matrix.STACK_AncShiftNode = matrix.STACK_AncShiftNode, matrix.STACK_AncShiftValue = matrix.STACK_AncShiftValue, 
                                          vector.Depth = runif(n = numeric.NumberTrainingReps, min = 0, max = 10),
                                          list.Rmatrix = list.Rmatrix)

#################################################################
# simulate under depth model with STACKED 01 (Depth + AncShift) #
#################################################################
Simulate.TraitTrain.depth_STACKED_01(numeric.NumberOfSpecies = length(handle.PrimatePhylogeny$tip.label), 
                                     handle.Phylogeny = handle.PrimatePhylogeny, 
                                     numeric.AncestralState = list.SimulationModelSettings[[1]]$vector.AncestralState[1], 
                                     numeric.depth = list.SimulationModelSettings[[1]]$vector.Depth[1],
                                     vector.STACK_AncShiftNode = list.SimulationModelSettings[[1]]$matrix.STACK_AncShiftNode[,1], 
                                     vector.STACK_AncShiftValue = list.SimulationModelSettings[[1]]$matrix.STACK_AncShiftValue[,1],
                                     matrix.R =  list.SimulationModelSettings[[1]]$list.Rmatrix[[1]])
}
