% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.TraitTrain.trend_00.R
\name{Simulate.TraitTrain.trend_00}
\alias{Simulate.TraitTrain.trend_00}
\title{Simulate.TraitTrain.trend_00: Function to simulate trait data under the trend model}
\usage{
Simulate.TraitTrain.trend_00(
  handle.Phylogeny,
  numeric.Sig2,
  numeric.AncestralState,
  numeric.slope,
  matrix.R
)
}
\arguments{
\item{handle.Phylogeny}{Phylogeny used to simulate training data under trend model}

\item{numeric.Sig2}{Numeric Value of the sigsg (evolutionary rate) parmater of the trend model}

\item{numeric.AncestralState}{Numerical value of the z0 (ancestral state) parameter of the trend model}

\item{matrix.R}{Matrix specifying the among trait covariance for p traits. Can also be a single value (1) for a single trait}
}
\value{
matrix.TraitData Matrix of simulated trait data. Species are rows, traits are columns
}
\description{
This function returns a Matrix of simulated trait data. Species are rows, traits are columns in matrix.TraitData
}
\examples{
#############
# Reinstall #
#############
rm(list = ls()); setwd("~/OneDrive/ARCHIVE/_Software/"); uninstall("TraitTrainR/"); setwd("./TraitTrainR/"); document(); setwd("~/OneDrive/ARCHIVE/_Software/"); install("TraitTrainR")

#####################
# load dependencies #
#####################
library(geiger); library(phytools); library(TraitTrainR)

#########################
# get example phylogeny #
#########################
handle.PrimatePhylogeny <- read.tree(text = "((((((((human: 6, chimp:6): 1, gorilla: 7): 7, orangutan: 14): 11, macaque: 25): 64, mouse: 89): 91, opossum: 180): 20, platypus: 200): 110, chicken: 310);")

#############################
# Simulation Model Settings #
#############################
list.SimulationModelSettings <- list() # define an empty model list

##########################
# SET SIMULATION MODEL 1 #
##########################
numeric.NumberTrainingReps <- 10 # same number of replicates for all models in list.SimulationModelSettings
list.Rmatrix <- list(); for (i in 1:numeric.NumberTrainingReps){list.Rmatrix[[i]] <- matrix(c(1, 0, 0, 0, 1, 0, 0, 0, 1), nrow = 3, ncol = 3)} # three traits. Different rates for different traits can be specified here. 
#list.Rmatrix <- list(); for (i in 1:numeric.NumberTrainingReps){list.Rmatrix[[i]] <- 1} # single trait

#############################
# append to simulation list #
#############################
list.SimulationModelSettings[[1]] <- list(string.SimulationModel = "trend", 
                                          vector.Sig2 = rexp(n = numeric.NumberTrainingReps, rate = 1), 
                                          vector.AncestralState = runif(n = numeric.NumberTrainingReps, min = 0, max = 100), 
                                          vector.slope = rexp(n = numeric.NumberTrainingReps, rate = 1),
                                          list.Rmatrix = list.Rmatrix)

##########################################################################
# simulate under trend model with STACKED 11 (trend + lrates + AncShift) #
##########################################################################
Simulate.TraitTrain.trend_00(handle.Phylogeny = handle.PrimatePhylogeny, 
                             numeric.Sig2 = list.SimulationModelSettings[[1]]$vector.Sig2[1], 
                             numeric.AncestralState = list.SimulationModelSettings[[1]]$vector.AncestralState[1], 
                             numeric.slope = list.SimulationModelSettings[[1]]$vector.slope[1], 
                             matrix.R = list.SimulationModelSettings[[1]]$list.Rmatrix[[1]])
}
